refactor(GenericContainer): extract network mode logic into NetworkModeSetting trait
The network mode logic has been refactored into a new trait,
NetworkModeSetting, to improve modularity and reusability.
This change removes the network mode properties and methods
from the GenericContainer class and encapsulates them within
the trait. This allows for cleaner code and easier maintenance.
Additionally, a new test class, NetworkModeSettingTest, has been
added to ensure the functionality of the NetworkModeSetting trait.
The BaseCommand trait has been updated to handle objects with
a __toString method, enhancing its flexibility.
