+++ tests/Unit/Containers/GenericContainerTest.php
@@ -69,17 +69,6 @@ class GenericContainerTest extends TestCase
         $this->assertSame("Hello, World!\n", $instance->getOutput());
     }
 
-    public function testStartWithNetworkMode()
-    {
-        $container = (new GenericContainer('alpine:latest'))
-            ->withNetworkMode('none')
-            ->withCommands(['sh', '-c', 'ls /sys/class/net']);
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $instance = $container->start();
-
-        $this->assertFalse(strpos($instance->getOutput(), 'eth0'));
-    }
-
     public function testStartWithNetworkAliases()
     {
         $instance = Testcontainers::run(DinD::class);
